creates new WidgetMode module
