+++ src/modules/index.ts
@@ -2,3 +2,4 @@ export * from "./colors";
 export * from "./date";
 export { http } from "./http";
 export * from "./asyncInitializer";
+export * from "./WidgetMode";
